Move restraint plots to separate code
